Fixed Long Data

+++ src/main/java/me/itsmas/timedaccess/util/UtilTime.java
@@ -10,7 +10,7 @@ public final class UtilTime
     {
         try
         {
-            return Integer.parseInt(input);
+            return Long.parseLong(input);
         }
         catch (IllegalArgumentException ex)
         {
